relocate middleware folder
