+++ packages/dappmanager/src/api/middlewares/ethForward/utils/decodeContent.ts
