+++ packages/dappmanager/src/api/middlewares/ethForward/utils/isEmpty.ts
