+++ packages/dappmanager/src/api/middlewares/ethForward/views/base.ts
