+++ packages/dappmanager/src/api/middlewares/viewsCounter/index.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import * as db from "../../db";
+import * as db from "../../../db";
 
 let counter = 0;
